# openapi_schema.json
{"openapi": "3.1.0", "info": {"title": "FastAPI", "version": "0.1.0"}, "paths": {"/": {"get": {"summary": "Read Root", "operationId": "read_root__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/token/": {"post": {"summary": "Login For Access Token", "operationId": "login_for_access_token_token__post", "requestBody": {"content": {"application/x-www-form-urlencoded": {"schema": {"$ref": "#/components/schemas/Body_login_for_access_token_token__post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/TokenReturn"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/{user_id}": {"get": {"summary": "Get User By Id", "operationId": "get_user_by_id_users__user_id__get", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserPublic"}}}}, "404": {"description": "User not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NotFoundResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "patch": {"summary": "Patch User", "operationId": "patch_user_users__user_id__patch", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserUpdate"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserPublic"}}}}, "404": {"description": "User not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NotFoundResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"summary": "Delete User", "operationId": "delete_user_users__user_id__delete", "parameters": [{"name": "user_id", "in": "path", "required": true, "schema": {"type": "string", "title": "User Id"}}], "responses": {"204": {"description": "Successful Response"}, "404": {"description": "User not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NotFoundResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/username/{username}": {"get": {"summary": "Get User By Username", "operationId": "get_user_by_username_users_username__username__get", "parameters": [{"name": "username", "in": "path", "required": true, "schema": {"type": "string", "minLength": 3, "maxLength": 20, "title": "Username"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserPublic"}}}}, "404": {"description": "User not found", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/NotFoundResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/": {"post": {"summary": "Post User", "operationId": "post_user_users__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserCreate"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserPublic"}}}}, "409": {"description": "User already exists", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DetailOnlyResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/available/username/{username}/": {"get": {"summary": "User Username Available", "operationId": "user_username_available_users_available_username__username___get", "parameters": [{"name": "username", "in": "path", "required": true, "schema": {"type": "string", "minLength": 3, "maxLength": 20, "title": "Username"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ItemAvailableResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/users/available/email/{email}/": {"get": {"summary": "User Username Exists", "operationId": "user_username_exists_users_available_email__email___get", "parameters": [{"name": "email", "in": "path", "required": true, "schema": {"type": "string", "format": "email", "maxLength": 254, "title": "Email"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ItemAvailableResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/pages/profile/": {"get": {"summary": "Get Pages Profile", "operationId": "get_pages_profile_pages_profile__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "expiry_timedelta", "in": "query", "required": false, "schema": {"type": "string", "format": "duration", "default": "P7D", "title": "Expiry Timedelta"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PagesProfileResponse"}}}}, "401": {"description": "Unauthorized", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DetailOnlyResponse"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/pages/home/": {"get": {"summary": "Get Pages Home", "operationId": "get_pages_home_pages_home__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "expiry_timedelta", "in": "query", "required": false, "schema": {"type": "string", "format": "duration", "default": "P7D", "title": "Expiry Timedelta"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Body_login_for_access_token_token__post": {"properties": {"grant_type": {"anyOf": [{"type": "string", "pattern": "password"}, {"type": "null"}], "title": "Grant Type"}, "username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}, "scope": {"type": "string", "title": "Scope", "default": ""}, "client_id": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Id"}, "client_secret": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Client Secret"}}, "type": "object", "required": ["username", "password"], "title": "Body_login_for_access_token_token__post"}, "DetailOnlyResponse": {"properties": {"detail": {"type": "string", "title": "Detail"}}, "type": "object", "required": ["detail"], "title": "DetailOnlyResponse"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "ItemAvailableResponse": {"properties": {"available": {"type": "boolean", "title": "Available"}}, "type": "object", "required": ["available"], "title": "ItemAvailableResponse"}, "NotFoundResponse": {"properties": {"detail": {"type": "string", "title": "Detail"}}, "type": "object", "required": ["detail"], "title": "NotFoundResponse"}, "PagesProfileResponse": {"properties": {"user": {"anyOf": [{"$ref": "#/components/schemas/UserPublic"}, {"type": "null"}]}}, "type": "object", "title": "PagesProfileResponse"}, "Token": {"properties": {"access_token": {"type": "string", "title": "Access Token"}, "token_type": {"type": "string", "title": "Token Type"}}, "type": "object", "required": ["access_token", "token_type"], "title": "Token"}, "TokenReturn": {"properties": {"token": {"$ref": "#/components/schemas/Token"}, "user": {"$ref": "#/components/schemas/UserPublic"}}, "type": "object", "required": ["token", "user"], "title": "TokenReturn"}, "UserCreate": {"properties": {"username": {"type": "string", "maxLength": 20, "minLength": 3, "title": "Username"}, "email": {"type": "string", "maxLength": 254, "minLength": 1, "format": "email", "title": "Email"}, "password": {"type": "string", "maxLength": 64, "minLength": 1, "title": "Password"}}, "type": "object", "required": ["username", "email", "password"], "title": "UserCreate"}, "UserPublic": {"properties": {"id": {"type": "string", "title": "Id"}, "username": {"type": "string", "maxLength": 20, "minLength": 3, "title": "Username"}}, "type": "object", "required": ["id", "username"], "title": "UserPublic"}, "UserUpdate": {"properties": {"username": {"anyOf": [{"type": "string", "maxLength": 20, "minLength": 3}, {"type": "null"}], "title": "Username"}, "email": {"anyOf": [{"type": "string", "maxLength": 254, "minLength": 1, "format": "email"}, {"type": "null"}], "title": "Email"}, "password": {"anyOf": [{"type": "string", "maxLength": 64, "minLength": 1}, {"type": "null"}], "title": "Password"}}, "type": "object", "title": "UserUpdate"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "token"}}}}}}